Create 1.c
In the *.c file it should define a function and call it in the main()
The function should promote a string “Enter 10 integers: ” to the screen. 
The user will enter 10 integers. 
The function should calculate and return the average of the 10 inputs. 
The main() should show the average in a newline on the screen once the previous function returns a value.